refactor: gather the initialization methods in the subpackage initializers
initialization.py splits into initializers/methods.py (protocol, registry,
init_prior), _base.py (InitResult, candidate selection), _search.py (the base
class of the search methods), exact.py, sampling.py (the box samplers and
uniform), warmstart.py and cmaes.py. warm_start moves into
initializers/warmstart.py next to its strategy class.

Breaking: el.initialization is now el.initializers, with el.initializers.uniform
for the box; el.initialization.hyper_names is el.optimizers.search.hyper_names;
el.warmstart.warm_start is el.initializers.warmstart.warm_start.
