+++ tests/integration/test_invertiblenetwork.py
@@ -98,7 +98,7 @@ def test_parametric_prior():
     initializer = el.initializer(
         method="random",
         iterations=32,
-        distribution=el.initialization.uniform(radius=2.0, mean=0.0),
+        distribution=el.initializers.uniform(radius=2.0, mean=0.0),
     )
 
     eliobj = el.Elicit(
@@ -204,7 +204,7 @@ def test_invertible_network():
     initializer = el.initializer(
         method="random",
         iterations=32,
-        distribution=el.initialization.uniform(radius=2.0, mean=0.0),
+        distribution=el.initializers.uniform(radius=2.0, mean=0.0),
     )
 
     eliobj = el.Elicit(
